Move checklist component
