+++ src/parser/core/modules/Checklist/Checklist.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {i18nMark} from '@lingui/react'
 
 import Rule from './Rule'
-import ChecklistComponent from 'components/modules/Checklist'
+import ChecklistComponent from './Component'
 import Module, {DISPLAY_MODE} from 'parser/core/Module'
 import DISPLAY_ORDER from '../DISPLAY_ORDER'
