+++ src/parser/core/modules/Checklist/Component.js
